requirements.txt
SQLAlchemy==2.0.23
SQLAlchemy[asyncio]
uvicorn>=0.24.0
starlette>=0.32.0
strawberry-graphql>=0.215.0
python-decouple>=3.8